# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Account API.",
        "title": "Swagger Account API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:18888",
    "basePath": "/api/v1",
    "paths": {
        "/protected/project/creater_id": {
            "get": {
                "description": "Get project by createrId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "Get project by createrId",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}